// src/modules/persons/responses/types/IDreamResponse.ts
import { type IComment } from '@modules/projects/infra/repositories/entities/IComment'

import { type IPersonInObject } from './IPersonInObject'

export interface IDreamResponse {
  id: string
  infos: {
    title: string
    description: string
    createdAt: Date
  }
  collections: {
    referencesIt: {
      itensLength: number
      itens: IPersonInObject[]
    }
    comment: {
      itensLength: number
      itens: IComment[]
    }
  }
}